+++ lib/Musical_MPR_Sensor/Musical_MPR_Sensor.h
@@ -1,7 +1,9 @@
 #ifndef MUSICAL_MPR_SENSOR_H 
 #define MUSICAL_MPR_SENSOR_H
 
+#include "Data_File.h"
+
 bool mpr_Setup();
-int mpr_CheckTouch();
+uint8_t mpr_CheckTouch();
 
 #endif
